# openapi.yaml
openapi: 3.0.3
info:
  title: Purrfect API
  version: 1.0.0
  description: >
    Secure cat marketplace backend with escrow milestone settlement,
    veterinary inspection window, and dispute moderation.
servers:
  - url: https://api.purrfect.kz
    description: Production
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /listings/{listingId}/media:
    post:
      tags: [listings]
      summary: Add media to listing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mediaType, storageUrl, checksumSha256]
              properties:
                mediaType:
                  type: string
                  enum: [PHOTO, VIDEO]
                storageUrl:
                  type: string
                  format: uri
                checksumSha256:
                  type: string
      responses:
        '201': { description: Media added }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [listings]
      summary: List listing media
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200': { description: Listing media list }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /listings/{listingId}/documents:
    post:
      tags: [listings]
      summary: Add document to listing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '201': { description: Document added }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [listings]
      summary: List listing documents
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
      responses:
        '200': { description: Listing documents list }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /listings/{listingId}/documents/{documentId}/verify:
    post:
      tags: [moderation]
      summary: Verify listing document as moderator
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
        - in: path
          name: documentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201': { description: Document verification created }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /listings/{listingId}/media/{mediaId}:
    delete:
      tags: [listings]
      summary: Remove a media item from a listing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
        - in: path
          name: mediaId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Media removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /listings/{listingId}/documents/{documentId}:
    delete:
      tags: [listings]
      summary: Remove a document from a listing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ListingIdParam'
        - in: path
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Document removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ListingIdParam:
      in: path
      name: listingId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: BAD_REQUEST
            message: Invalid query parameter value
    Unauthorized:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Access token is invalid or expired
    Forbidden:
      description: Authenticated but missing required role/permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Admin role is required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Listing not found
    Conflict:
      description: Business state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: Order is already in disputed state
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: inspection report URL is required
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Unexpected server error
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
...